requirements.txt
Django==5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg==3.1.14

